Move the `tests::snapshot` submodule into its own file
Having these tests in an inline module made it harder to navigate both the
snapshot tests and the surrounding module.
